requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
feedparser>=6.0.11
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
msal>=1.31.0
pytest>=8.0.0
pytest-mock>=3.14.0
pytest-asyncio>=0.24.0
